pyproject.toml
[tool.poetry]
name = "pdfconv"
version = "0.1.0"
description = ""
authors = ["MPHRS <serafimin2002@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fpdf = "^1.7.2"
pypdf2 = "^3.0.1"
reportlab = "^3.6.13"
chardet = "^5.1.0"
docx2pdf = "^0.1.8"
python-pptx = "^0.6.21"
openpyxl = "^3.1.2"
img2pdf = "^0.4.4"
comtypes = "^1.1.14"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
